projects/avdoulou/requirements.txt
# Core Dependencies
python-telegram-bot==21.10
yt-dlp==2024.12.6
pydantic==2.10.4
pydantic-settings==2.6.1
aiohttp==3.11.11
aiofiles==24.1.0

# Testing Dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
